requirements.txt
absl-py==2.1.0
asttokens==2.4.1
attrs==23.2.0
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.10.0
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
filelock==3.9.0
flatbuffers==24.3.25
fonttools==4.50.0
fsspec==2023.4.0
gitdb==4.0.11
GitPython==3.1.43
idna==2.10
ipykernel==6.29.4
ipython==8.23.0
jax==0.4.26
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.6
mediapipe==0.10.11
ml-dtypes==0.4.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.8.0.74
opt-einsum==3.3.0
packaging==24.0
pandas==2.2.1
parso==0.8.3
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
pycparser==2.22
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
requests-toolbelt==1.0.0
roboflow==1.1.26
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sounddevice==0.4.6
stack-data==0.6.3
sympy==1.12
thop==0.1.1.post2209072238
torch==2.2.2+cu121
torchaudio==2.2.2+cu121
torchvision==0.17.2+cu121
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.8.0
tzdata==2024.1
ultralytics==8.1.42
urllib3==2.2.1
wcwidth==0.2.13
